Rewrite article detection
The previous version performed article detection in a needlessly
convoluted way, with lots of backwards-seeking. This was due, in part,
to the fact that there is no 'Front Page' section and possibly to the
fact that I did not know that I could search for a list of CSS classes.